Sensor File
Adds a file for sensor. Adds cases for testing
