+++ divider.py
@@ -1,8 +1,6 @@
 from config import *
 from surface import *
 from viewwindow import *
-
-
 class divider:
     def __init__(self, x, y):
         self.x = x
@@ -15,5 +13,6 @@ class divider:
         surface = SingletonSurface.getInstance().surface
         viewwindow = SingletonViewWindow.get_instance()
         r = viewwindow.transform_rect(self.x, self.y, self.w, self.h)
-
+       
         pygame.draw.rect(surface, self.color, r)
+
