+++ halter.py
@@ -1,8 +1,8 @@
 from init import *
 from helper_functions import *
 
-
 class Halter:
+
     def __init__(self):
         self.starttime = pygame.time.get_ticks()
 
@@ -18,3 +18,4 @@ class Halter:
         delay = get_delay(self.starttime)
         if delay > 0:
             pygame.time.delay(delay)
+
