+++ images/resize.py
@@ -1,7 +1,7 @@
 import os
-
-filename = "Mini_van"
+filename = 'Mini_van'
 width = 46
-os.system("convert -resize %s %s.png %s_small.png" % (width, filename, filename))
+os.system('convert -resize %s %s.png %s_small.png' % \
+          (width, filename, filename))
 
-# convert example.png -resize 200 example.png
+#convert example.png -resize 200 example.png
