+++ images/small/Car/rotate.py
@@ -1,8 +1,6 @@
 import os
 
-filename = "Car"
+filename = 'Car'
 for i in range(0, 360, 10):
-    os.system(
-        "convert -rotate %s -background 'rgba(0,0,0,0)' %s.png %s-%s.png"
-        % (i, filename, filename, i)
-    )
+    os.system("convert -rotate %s -background 'rgba(0,0,0,0)' %s.png %s-%s.png" % (i, filename, filename, i))
+
