+++ main.py
@@ -1,32 +1,4 @@
 from factory import *
+from basecaseMain import *
 
-carmodel, carcontrol, carview, audimodel, audicontrol, audiview, roadmodel, roadcontrol, roadview, viewwindow, halter, WorldObjects, sensorView, clock = (
-    factory()
-)
-while 1:
 
-    halter.update()
-    halter.run()
-
-    carcontrol.run()
-    viewwindow.run()
-    roadcontrol.run(WorldObjects)
-    audicontrol.run()
-
-    ########################################
-
-    roadview.run()
-    carmodel.sensor.detectWorld(carmodel, WorldObjects)
-    sensorView.run()
-    carview.run()
-    audiview.run()
-
-    clock.run()
-
-    pygame.display.flip()
-
-    # create an object called FramerateControl. Replace the following codes with
-    # frameratecontrol.run()
-
-    # delaying
-    halter.delay()
