+++ vehicle.py
@@ -11,127 +11,15 @@ from time import time
 
 
 def get_center(obj):
-    x = obj.x - (obj.w / (PIXEL)) / 2
+    x = obj.x + (obj.w / (PIXEL)) / 2
     y = obj.y - (obj.h / PIXEL) / 2
     return x, y
 
 
-# Sensor Class for the Car
-class Sensor:
-    def __init__(self, w, h):
-        self.w = w
-        self.h = h
-        # self.rect = pygame.Rect(x, y, w, h)
-        surface = SingletonSurface.getInstance().surface
-
-    def get_distance(self, audi, obj):
-        x1, y1 = get_center(audi)
-        x2, y2 = get_center(obj)
-        return math.sqrt(abs((pow(x1 - x2, 2) + pow(y1 - y2, 2))))
-
-    def detectWorld(self, audi, WorldObjects):
-        viewwindow = SingletonViewWindow.get_instance()
-        surface = SingletonSurface.getInstance().surface
-        nearby_objects = []
-        self.nearbyObjects = []
-
-        #############################################################
-        # Detects the grass for the audi
-        if audi.x < ROAD_X + (LANE_WIDTH):
-            # Detects the Grass on the Left
-            nearby_objects.append(("GRASS", ROAD_X, audi.y))
-            self.nearbyObjects.append(
-                (ROAD_X, audi.y - (audi.h / PIXEL), 1 / PIXEL, audi.h / PIXEL)
-            )
-
-        if audi.x > ROAD_X + (LANE_WIDTH * (NUM_LANES - 1)):
-            # Detects the Grass on the Right when in range
-            nearby_objects.append(("GRASS", ROAD_X + (LANE_WIDTH * NUM_LANES), audi.y))
-            self.nearbyObjects.append(
-                (
-                    ROAD_X + (LANE_WIDTH * NUM_LANES),
-                    audi.y - (audi.h / PIXEL),
-                    1 / PIXEL,
-                    audi.h / PIXEL,
-                )
-            )
-
-        if audi.x < ROAD_X or audi.x > ROAD_END:
-            print(
-                "<Car,",
-                round(audi.x),
-                ", ",
-                round(audi.y),
-                ", ",
-                audi.w,
-                ", ",
-                audi.h,
-                "> collides with GRASS",
-            )
-            nearby_objects.append(("GRASS", ROAD_X, audi.y, True))
-            # sys.exit()
-
-        #########################################################################################
-        # This now senses the other objects in the world and adds them to the nearby objects list
-        for obj in WorldObjects:
-            # Try to move the if-else using inheritance.
-            # Computes the distance
-
-            if isinstance(obj, VehicleModel):
-                if (
-                    self.get_distance(audi, obj) < MAX_RADIUS
-                    and self.get_distance(audi, obj) > 0
-                ):
-                    # Sensor code for cars
-                    rect = (obj.x, obj.y, obj.w / PIXEL, obj.h / PIXEL)
-                    mrect = (audi.x, audi.y, audi.w / PIXEL, audi.h / PIXEL)
-                    if colliderect(mrect, rect):
-
-                        # Return label and rectangle of object and let the sim decide if it needs to stop
-                        nearby_objects.append(("CAR", obj.x, obj.y, True))
-                        print(
-                            "<Car,",
-                            round(audi.x),
-                            ", ",
-                            round(audi.y),
-                            ", ",
-                            audi.w,
-                            ", ",
-                            audi.h,
-                            "> collides with CAR",
-                        )
-                        # sys.exit()
-
-                    # Adds the car object to sensed list
-                    self.nearbyObjects.append(
-                        (obj.x, obj.y - obj.h / (PIXEL), obj.w / PIXEL, obj.h / PIXEL)
-                    )
-                    nearby_objects.append(("CAR", obj.x, obj.y, False))
-
-            elif isinstance(obj, RoadModel):
-                for div in obj.dividers:
-                    if self.get_distance(audi, div) <= MAX_RADIUS:
-                        self.nearbyObjects.append((div.x, div.y, div.w, div.h))
-
-
-####
-
-
-class SensorView:
-    def __init__(self, Sensor):
-        self.sensor = Sensor
-
-    def run(self):
-        surface = SingletonSurface.getInstance().surface
-        viewwindow = SingletonViewWindow.get_instance()
-        for obj in self.sensor.nearbyObjects:
-            rect = viewwindow.transform_rect(obj[0], obj[1], obj[2], obj[3])
-            pygame.draw.rect(surface, RED, rect, 2)
-
 
 class VehicleModel:
     def __init__(
-        self, image, position, speed, accelaration, target_speed, max_speed, car_type
+            self, image, position, speed, accelaration, target_speed, max_speed, car_type, has_sensor = False
     ):
         self.image = image
         self.position = position
@@ -144,6 +32,10 @@ class VehicleModel:
         self.sensor = Sensor(self.w, self.h)
         self.car_type = car_type
         self.last_time = time()
+        if has_sensor:
+            self.sensor = Sensor(self.w, self.h)
+        else:
+            self.sensor = None
 
     def get_x(self):
         return self.position[0]
@@ -183,7 +75,7 @@ class VehicleModel:
     def set_accelaration_depression(self, accelaration_depression):
         self.accelaration_depression = accelaration_depression
 
-
+from sensor import *
 class VehicleControlUser:
     def __init__(self, vehiclemodel):
         self.vehiclemodel = vehiclemodel
@@ -226,6 +118,10 @@ class VehicleControlRandom:
 class VehicleView:
     def __init__(self, vehiclemodel):
         self.vehiclemodel = vehiclemodel
+        if vehiclemodel.sensor == None:
+            self.sensorView = None
+        else:
+            self.sensorView = SensorView(vehiclemodel.sensor)
 
     def run(self):
         x, y = self.vehiclemodel.position
@@ -234,4 +130,3 @@ class VehicleView:
         x, y = viewwindow.transform(x, y)
 
         surface.blit(self.vehiclemodel.image, (x, y))
-        # draw bounding box
